Add easter egg image viewer triggered by logo click
Introduces an easter egg feature that displays a random image from a predefined set when the site logo is clicked. Updates CSS for the logo to indicate interactivity and modifies image viewer logic to support this mode, disabling navigation controls and swipe gestures when an easter egg image is shown.